Fixed some logical errors in the Server code and added complete createRoom and joinRoom functionality.

+++ TriviaProjectNadavAndOren/TriviaProjectNadavAndOren/SqliteDatabase.hpp
@@ -194,7 +194,7 @@ public:
 
         float avgAnswerTime = 0.0;
         if (sqlite3_step(stmt) == SQLITE_ROW) {
-            avgAnswerTime = sqlite3_column_double(stmt, 0);
+            avgAnswerTime = (float)sqlite3_column_double(stmt, 0);
         }
 
         sqlite3_finalize(stmt);
